fixed BankAccountValidation
